separate content nav creation into utils
